+++ src/components/CardCvcNumber/CardCvcNumber.tsx
@@ -1,7 +1,7 @@
-import { useState } from "react";
 import NumberInput from "../Input/CardNumberInput";
 import InputErrorMessage from "../Input/InputErrorMessage";
 import InputText from "../InputText/InputText";
+import useCvcInputHandler from "../../hooks/useCvcNumber/useCvcInputHandler";
 
 interface CardCvcNumberProps {
   setCompleted: React.Dispatch<React.SetStateAction<boolean>>;
@@ -14,35 +14,14 @@ const CARD_CVC_NUMBER = {
   PLACEHOLDER: "123",
 } as const;
 
-const CVC_RULE = {
-  INVALID_LENGTH_ERROR: "CVC 번호는 3자리 숫자여야 합니다.",
-  MAX_LENGTH: 3,
-} as const;
-
 export default function CardCvcNumber({
   setCompleted,
   onComplete,
 }: CardCvcNumberProps) {
-  const [numbers, setNumbers] = useState("");
-  const [error, setError] = useState("");
-
-  const handleCardCvcNumberChange = (value: string) => {
-    if (value.length > CVC_RULE.MAX_LENGTH) return;
-
-    const validationError =
-      value.length < CVC_RULE.MAX_LENGTH ? CVC_RULE.INVALID_LENGTH_ERROR : "";
-
-    setNumbers(value);
-    setError(validationError);
-
-    const isComplete =
-      value.length === CVC_RULE.MAX_LENGTH && validationError === "";
-    setCompleted(isComplete);
-
-    if (isComplete) {
-      onComplete();
-    }
-  };
+  const { handleCardCvcNumberChange, numbers, error } = useCvcInputHandler(
+    setCompleted,
+    onComplete
+  );
 
   return (
     <section>
